requirements.txt
asgiref==3.4.1
Django==3.2.5
et-xmlfile==1.1.0
gensim==4.0.1
hazm==0.7.0
joblib==1.0.1
libwapiti==0.2.1
nltk==3.3
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
psycopg2==2.9.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
redis==3.5.3
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
smart-open==5.1.0
sqlparse==0.4.1
threadpoolctl==2.2.0
tqdm==4.62.2

djangorestframework~=3.12.4